Keep current #53 * Implement keep current option to increase current when cadence gets lower.

+++ src/firmware/main.c
@@ -55,7 +55,7 @@ void main(void)
 	sensors_init();
 
 	speed_sensor_set_signals_per_rpm(g_config.speed_sensor_signals);
-	pas_set_stop_delay(g_config.pas_stop_delay_x100s * 10);
+	pas_set_stop_delay((uint16_t)g_config.pas_stop_delay_x100s * 10);
 
 	battery_init();
 	throttle_init(g_config.throttle_start_voltage_mv, g_config.throttle_end_voltage_mv);
